LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/location-assign-expected.txt



FAIL Replace before load, triggered by location.assign() assert_equals: expected "http://localhost:8800/common/blank.html?thereplacement" but got "http://localhost:8800/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/resources/code-injector.html?pipe=sub(none)&code=location.assign(%60%2Fcommon%2Fblank.html%3Fthereplacement%60)%3B"
